Respect profile option to select multibuild
If the obs project is setup to use multibuild the plugin
used the first entry of the _multibuild file as the profile.
However this first entry can point to a different build
architecture and could be not the best choice for the
current host. Therefore if the --profile option is given
this value should have a higher priority for the checkout